+++ projects/node-api-test/src/index.js
@@ -37,6 +37,8 @@ if (!user) { throw new Error('GRAPHISTRY_USER environment variable not set'); }
 const password = process.env.GRAPHISTRY_PASSWORD;
 if (!password) { throw new Error('GRAPHISTRY_PASSWORD environment variable not set'); }
 
+const org = process.env.GRAPHISTRY_ORG;
+
 const protocol = process.env.GRAPHISTRY_PROTOCOL || 'https';
 const host = process.env.GRAPHISTRY_HOST || 'hub.graphistry.com';
 
@@ -44,7 +46,7 @@ const host = process.env.GRAPHISTRY_HOST || 'hub.graphistry.com';
 
 if (false) {
   
-    const client = new Client(user, password, protocol, host);
+    const client = new Client(user, password, org, protocol, host);
 
     const edgesFile = new EdgeFile(edges);
     const nodesFile = new NodeFile(nodes);  // optional
@@ -63,7 +65,7 @@ if (false) {
 
 } else if (false) {
 
-    const client = new Client(user, password, protocol, host);
+    const client = new Client(user, password, org, protocol, host);
 
     const edgesFile = new EdgeFile(edges);
     const nodesFile = new NodeFile(nodes);  // optional
@@ -81,7 +83,7 @@ if (false) {
 
 } else if (false) {
 
-    const client = new Client(user, password, protocol, host);
+    const client = new Client(user, password, org, protocol, host);
 
     const edgesFile = new EdgeFile(
         edgesRows,
@@ -182,7 +184,7 @@ if (false) {
 
 } else if (false) {
 
-    const client = new Client(user, password, protocol, host);
+    const client = new Client(user, password, org, protocol, host);
     const edgesFile = new EdgeFile(edges);
     const nodesFile = new NodeFile(nodes);  // optional
 
@@ -206,7 +208,7 @@ if (false) {
     //convert edges to apache-arrow table
     const edgesArr = tableFromArrays(edges);
     const nodesArr = tableFromArrays(nodes);
-    const client = new Client(user, password, protocol, host);
+    const client = new Client(user, password, org, protocol, host);
 
     function arrToUint8Array(arr) {
         const ui8 = tableToIPC(arr, 'file');
@@ -233,7 +235,7 @@ if (false) {
     //convert edges to apache-arrow table
     const edgesArr = tableFromArrays(edges);
     const nodesArr = tableFromArrays(nodes);
-    const client = new Client(user, password, protocol, host);
+    const client = new Client(user, password, org, protocol, host);
 
     function arrToUint8Array(arr) {
         const ui8 = tableToIPC(arr, 'file');