+++ CHANGELOG.md
@@ -3,12 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## Development
+## 4.6.0
 
 ### Added
 
 - **js-upload-api**: CommonJS support with output binaries at `dist/cjs`
 - **node-api**: CommonJS support with output binaries at `dist/cjs`
+- **all**: Optional orgs login
+
+### Docs
+
+- **all**: Arrow uploads and login modes
 
 ## 4.5.0 - 2023-03-23